move tests from programmes page to sessions page
